# TradingWebApp.Server/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "TradingWebApp.Server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Quote/Lookup": {
      "post": {
        "tags": [
          "Quote"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Stock/Read": {
      "get": {
        "tags": [
          "Stock"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Transaction/Buy": {
      "post": {
        "tags": [
          "Transaction"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BuyModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Transaction/Sell": {
      "post": {
        "tags": [
          "Transaction"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BuyModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Transaction/GetTransactions/{userId}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/Register": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/Login": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/Logout": {
      "post": {
        "tags": [
          "User"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/Read": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userID",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "BuyModel": {
      "required": [
        "name",
        "numberOfShares",
        "shareValue",
        "symbol",
        "totalBookCost",
        "userID"
      ],
      "type": "object",
      "properties": {
        "userID": {
          "format": "int32",
          "type": "integer"
        },
        "symbol": {
          "minLength": 1,
          "type": "string"
        },
        "name": {
          "minLength": 1,
          "type": "string"
        },
        "numberOfShares": {
          "format": "int32",
          "type": "integer"
        },
        "totalBookCost": {
          "format": "double",
          "type": "number"
        },
        "shareValue": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "LoginModel": {
      "required": [
        "password",
        "username"
      ],
      "type": "object",
      "properties": {
        "username": {
          "minLength": 6,
          "type": "string"
        },
        "password": {
          "format": "password",
          "minLength": 6,
          "type": "string"
        }
      }
    },
    "RegisterModel": {
      "required": [
        "confirmPassword",
        "password",
        "username"
      ],
      "type": "object",
      "properties": {
        "username": {
          "minLength": 6,
          "type": "string"
        },
        "password": {
          "format": "password",
          "minLength": 6,
          "type": "string"
        },
        "confirmPassword": {
          "format": "password",
          "minLength": 1,
          "type": "string"
        }
      }
    }
  }
}